Add a new StatisticsError exception class
Also make an outer `with Env():` optional when using an opened
dataset as a context manager.

Resolves #2759
