+++ rasterio/transform.py
@@ -60,13 +60,13 @@ class TransformMethodsMixin:
         col : int
             Pixel column.
         z : float, optional
-            Height associated with coordinates. Primarily used for RPC based 
-            coordinate transformations. Ignored for affine based 
+            Height associated with coordinates. Primarily used for RPC based
+            coordinate transformations. Ignored for affine based
             transformations. Default: 0.
         offset : str, optional
             Determines if the returned coordinates are for the center of the
             pixel or for a corner.
-        transform_method: TransformMethod, optional 
+        transform_method: TransformMethod, optional
             The coordinate transformation method. Default: `TransformMethod.affine`.
         rpc_options: dict, optional
             Additional arguments passed to GDALCreateRPCTransformer
@@ -103,13 +103,13 @@ class TransformMethodsMixin:
         y : float
             y value in coordinate reference system
         z : float, optional
-            Height associated with coordinates. Primarily used for RPC based 
-            coordinate transformations. Ignored for affine based 
+            Height associated with coordinates. Primarily used for RPC based
+            coordinate transformations. Ignored for affine based
             transformations. Default: 0.
         op : function, optional (default: math.floor)
             Function to convert fractional pixels to whole numbers (floor,
             ceiling, round)
-        transform_method: TransformMethod, optional 
+        transform_method: TransformMethod, optional
             The coordinate transformation method. Default: `TransformMethod.affine`.
         rpc_options: dict, optional
             Additional arguments passed to GDALCreateRPCTransformer
@@ -210,20 +210,20 @@ def xy(transform, rows, cols, zs=None, offset='center', **rpc_options):
     The pixel's center is returned by default, but a corner can be returned
     by setting `offset` to one of `ul, ur, ll, lr`.
 
-    Supports affine, Ground Control Point (GCP), or Rational Polynomial 
+    Supports affine, Ground Control Point (GCP), or Rational Polynomial
     Coefficients (RPC) based coordinate transformations.
 
     Parameters
     ----------
     transform : Affine or sequence of GroundControlPoint or RPC
-        Transform suitable for input to AffineTransformer, GCPTransformer, or RPCTransformer. 
+        Transform suitable for input to AffineTransformer, GCPTransformer, or RPCTransformer.
     rows : list or int
         Pixel rows.
     cols : int or sequence of ints
         Pixel columns.
     zs : list or float, optional
-        Height associated with coordinates. Primarily used for RPC based 
-        coordinate transformations. Ignored for affine based 
+        Height associated with coordinates. Primarily used for RPC based
+        coordinate transformations. Ignored for affine based
         transformations. Default: 0.
     offset : str, optional
         Determines if the returned coordinates are for the center of the
@@ -256,8 +256,8 @@ def rowcol(transform, xs, ys, zs=None, op=math.floor, precision=None, **rpc_opti
     ys : list or float
         y values in coordinate reference system.
     zs : list or float, optional
-        Height associated with coordinates. Primarily used for RPC based 
-        coordinate transformations. Ignored for affine based 
+        Height associated with coordinates. Primarily used for RPC based
+        coordinate transformations. Ignored for affine based
         transformations. Default: 0.
     op : function
         Function to convert fractional pixels to whole numbers (floor, ceiling,
@@ -322,7 +322,7 @@ class TransformerBase():
     @staticmethod
     def _ensure_arr_input(xs, ys, zs=None):
         """Ensure all input coordinates are mapped to array-like objects
-        
+
         Raises
         ------
         TransformError
@@ -343,7 +343,7 @@ class TransformerBase():
     def __exit__(self, *args):
         self.close()
         self._env.close()
-    
+
     def rowcol(self, xs, ys, zs=None, op=math.floor, precision=None):
         """Get rows and cols coordinates given geographic coordinates.
 
@@ -352,8 +352,8 @@ class TransformerBase():
         xs, ys : float or list of float
             Geographic coordinates
         zs : float or list of float, optional
-            Height associated with coordinates. Primarily used for RPC based 
-            coordinate transformations. Ignored for affine based 
+            Height associated with coordinates. Primarily used for RPC based
+            coordinate transformations. Ignored for affine based
             transformations. Default: 0.
         op : function, optional (default: math.floor)
             Function to convert fractional pixels to whole numbers (floor,
@@ -380,7 +380,7 @@ class TransformerBase():
 
         AS_ARR = True if hasattr(xs, "__iter__") else False
         xs, ys, zs = self._ensure_arr_input(xs, ys, zs=zs)
-        
+
         try:
             new_cols, new_rows = self._transform(
                 xs, ys, zs, transform_direction=TransformDirection.reverse
@@ -402,8 +402,8 @@ class TransformerBase():
         rows, cols : int or list of int
             Image pixel coordinates
         zs : float or list of float, optional
-            Height associated with coordinates. Primarily used for RPC based 
-            coordinate transformations. Ignored for affine based 
+            Height associated with coordinates. Primarily used for RPC based
+            coordinate transformations. Ignored for affine based
             transformations. Default: 0.
         offset : str, optional
             Determines if the returned coordinates are for the center of the
@@ -420,7 +420,7 @@ class TransformerBase():
         """
         AS_ARR = True if hasattr(rows, "__iter__") else False
         rows, cols, zs = self._ensure_arr_input(rows, cols, zs=zs)
-        
+
         if offset == 'center':
             coff, roff = (0.5, 0.5)
         elif offset == 'ul':
@@ -433,7 +433,7 @@ class TransformerBase():
             coff, roff = (1, 1)
         else:
             raise TransformError("Invalid offset")
-        
+
         # shift input coordinates according to offset
         T = IDENTITY.translation(coff, roff)
         offset_rows = []
@@ -454,7 +454,7 @@ class TransformerBase():
                 return (new_xs, new_ys)
         except TypeError:
             raise TransformError("Invalid inputs")
-    
+
     def _transform(self, xs, ys, zs, transform_direction):
         raise NotImplementedError
 
@@ -469,11 +469,11 @@ class AffineTransformer(TransformerBase):
 
     def close(self):
         pass
-    
+
     def _transform(self, xs, ys, zs, transform_direction):
         resxs = []
         resys = []
-        
+
         if transform_direction is TransformDirection.forward:
             transform = self._transformer
         elif transform_direction is TransformDirection.reverse:
@@ -483,9 +483,9 @@ class AffineTransformer(TransformerBase):
             resx, resy = transform * (x, y)
             resxs.append(resx)
             resys.append(resy)
-        
+
         return (resxs, resys)
-    
+
     def __repr__(self):
         return "<AffineTransformer>"
 
@@ -497,7 +497,7 @@ class RPCTransformer(RPCTransformerBase, TransformerBase):
 
     Uses GDALCreateRPCTransformer and GDALRPCTransform for computations. Options
     for GDALCreateRPCTransformer may be passed using `rpc_options`.
-    Ensure that GDAL transformer objects are destroyed by calling `close()` 
+    Ensure that GDAL transformer objects are destroyed by calling `close()`
     method or using context manager interface.
 
     """
@@ -517,7 +517,7 @@ class GCPTransformer(GCPTransformerBase, TransformerBase):
     coordinate transformations.
 
     Uses GDALCreateGCPTransformer and GDALGCPTransform for computations.
-    Ensure that GDAL transformer objects are destroyed by calling `close()` 
+    Ensure that GDAL transformer objects are destroyed by calling `close()`
     method or using context manager interface.
 
     """
